implement Module::get_export
